backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-cloud-texttospeech==2.21.0
google-auth==2.36.0
requests==2.32.0
python-dotenv==1.0.1
